Repetier-Host: new package
FIXME: slic3r doesn't work here either; 1) because it's missing lots of
perl modules (I've already added them to nixpkgs, but not to this
expression) and 2) because slic3r doesn't work with perl 5.16...
----

Repetier-Host is software for controlling RepRap style 3D-printer like
Mendel, Darwin or Prusa mendel. Works with most firmware types. It is
optimized to work with Repetier-Firmware Other working firmware is
Sprinter, Teacup, Marlin and all compatible firmwares.

I'm not familiar enough with mono and how to reference assemblies, so
I'm using the prebuilt binary distribution of Repetier-Host for now.
Building from source is on the TODO list.

I have not tested this with a 3D-printer, only build and run tested.