WIP: nix-prefetch-git: bring back git index (with leaveDotGit)
IMPOSSIBLE: The git index contains really non-deterministic data:
  $ git ls-files --debug
  README.txt
    ctime: 1415112882:583529058
    mtime: 1415112882:583529058
    dev: 2049	ino: 4718886
    uid: 1000	gid: 100
    size: 6	flags: 0

  Timestamps can be faked (libfaketime) but dev, inode number
  (especially this!), uid/gid will be quite a "challenge"...

The git index contains timestamps. So far we have removed the index to
have deterministic output. But removing the index makes the tree
"dirty". Not good for build systems that use "git describe --dirty" to
generate a version string.

This patch changes nix-prefech-git (and fetchgit) to run git with
libfaketime (and time set to UNIX epoch). This way we can keep the index
and still have deterministic output.

Implementation comments:
* Use bash arrays in the wrappers to be able to pass FAKETIME
  environment variable, which contains whitespace, to the
  nix-prefetch-git wrapper.
* nix-prefetch-git is part of the nix-prefetch-scripts package. Only
  the *-git script is modified by this patch, as I don't know how fake
  time would affect the other fetchers.