v4l-utils/libv4l: add symlink for #include <linux/videodev.h>
(In addition to the existing one for #include <videodev.h>.)

When doing online searching for "videodev.h", most hits come up with the
"linux/" prefix (and in the context of "which package to install to
satisfy '#include <linux/videodev.h>'"). So this seems to be the
canonical location for that compatibility header.

I first hit this missing file when working on zbar, but in the end it
was patched to use V4L2 headers instead. But it seems other (old)
programs still assume that file location, based on the online search
results: flashcam, opencv