Revert "WIP: evolution-ews: remove patching of install dirs (hopefully fixes E + EDS + EWS "super" build)"
This reverts commit e418a62fd0abf36bbecfe65fa76b4d4557da2691.

The evolution-with-plugins build currently fails; it seems
evolution-ews is built even though it shouldn't (I only fetch
buildInputs and configureFlags from the evolution-ews attribute, I
didn't expect it to be built for that).

So try to unbreak by letting eolution-ews install into its own
directory. For the super-build, this is irrelevant.